# internal/handler/openapi.yaml
openapi: 3.1.0
info:
  title: ShieldNet Gateway Control Plane API
  description: |
    REST API for ShieldNet Gateway (SNG) — multi-tenant control plane.
    All operator endpoints are mounted under `/api/v1`. Auth is
    either a bearer JWT (operator console) or `X-SNG-API-Key`
    (machine-to-machine).
  version: "0.1.0"
  contact:
    name: ShieldNet Gateway Team
servers:
  - url: https://control.shieldnet.example/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local dev
security:
  - BearerJWT: []
  - APIKey: []
...
paths:
...
  /tenants/{tenant_id}/baselines:
    parameters:
      - $ref: "#/components/parameters/TenantPathID"
    get:
      tags: [baseline]
      operationId: listBaselineModels
      summary: List baseline statistical models for the tenant
      description: |
        Returns one row per (dimension, window) currently
        modelled. Each row carries Welford + EWMA state plus
        the configured z-score threshold used by the anomaly
        detector.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: page of baseline models
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/BaselineModel" }
                  next_cursor: { type: string }
  /tenants/{tenant_id}/baselines/{dimension}/{window_seconds}:
    parameters:
      - $ref: "#/components/parameters/TenantPathID"
      - name: dimension
        in: path
        required: true
        schema: { type: string }
      - name: window_seconds
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    get:
      tags: [baseline]
      operationId: getBaselineModel
      summary: Fetch a single baseline model
      responses:
        "200":
          description: baseline model
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaselineModel" }
        "404":
          description: model not found
  /tenants/{tenant_id}/baselines/{dimension}/{window_seconds}/threshold:
    parameters:
      - $ref: "#/components/parameters/TenantPathID"
      - name: dimension
        in: path
        required: true
        schema: { type: string }
      - name: window_seconds
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    put:
      tags: [baseline]
      operationId: updateBaselineThreshold
      summary: Override the z-score threshold for this baseline
      description: |
        Threshold updates take effect on the next anomaly
        evaluation; the in-flight observation pipeline is not
        interrupted. The Feedback service may later re-tune
        this value from operator feedback.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [z_threshold]
              properties:
                z_threshold:
                  type: number
                  format: double
                  minimum: 0
      responses:
        "200":
          description: updated model
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaselineModel" }
        "404":
          description: model not found
...
components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
    APIKey:
      type: apiKey
      in: header
      name: X-SNG-API-Key
  parameters:
...
    TenantPathID:
      in: path
      name: tenant_id
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    BaselineModel:
      type: object
      required: [tenant_id, dimension, window_seconds, mean, variance, ewma, z_threshold, samples, updated_at]
      properties:
        tenant_id: { type: string, format: uuid }
        dimension: { type: string }
        window_seconds: { type: integer, format: int64 }
        mean: { type: number, format: double }
        variance: { type: number, format: double }
        ewma: { type: number, format: double }
        ewma_variance: { type: number, format: double }
        z_threshold: { type: number, format: double }
        samples: { type: integer, format: int64 }
        last_observed_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...